+++ src/mds/MDCache.h
@@ -1599,26 +1599,6 @@ private:
   std::chrono::milliseconds quiesce_sleep;
 };
 
-class C_MDS_RetryRequest : public MDSInternalContext {
-  MDCache *cache;
-  MDRequestRef mdr;
- public:
-  C_MDS_RetryRequest(MDCache *c, const MDRequestRef& r) :
-    MDSInternalContext(c->mds), cache(c), mdr(r) {}
-  void finish(int r) override;
-};
-
-class CF_MDS_RetryRequestFactory : public MDSContextFactory {
-public:
-  CF_MDS_RetryRequestFactory(MDCache *cache, const MDRequestRef& mdr, bool dl) :
-    mdcache(cache), mdr(mdr), drop_locks(dl) {}
-  MDSContext *build() override;
-private:
-  MDCache *mdcache;
-  MDRequestRef mdr;
-  bool drop_locks;
-};
-
 /**
  * Only for contexts called back from an I/O completion
  *
